refactor: 로또프린터를 ui로 이동
