+++ src/main/java/lotto/service/LottoService.java
@@ -1,6 +1,8 @@
 package lotto.service;
 
 import lotto.domain.LottoCashier;
+import lotto.ui.LottoPrinter;
+import lotto.domain.Lottos;
 import lotto.domain.PurchaseAmount;
 import lotto.dto.LottoPurchaseResult;
